Improve abstraction of where condition evaluation
